requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
fastapi-users[sqlalchemy]==12.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic[email]==2.5.3
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
beautifulsoup4==4.12.3
requests==2.31.0
pypdf2==3.0.1
python-docx==1.1.0

# PDF Generation
reportlab==4.0.9

# LLM APIs
anthropic==0.18.1
openai==1.12.0

# Utilities
httpx==0.26.0
